Make emoji tooltips appear on every <Emoji/>
Previously, emoji tooltips were essentially tied to the emoji displayed
in the search results. This patch makes tooltips appear on every usage
of the Emoji component, which will help in the future.

Additionally, the horizontal tooltip positioning has been improved: it
no longer relies on the width of the emoji, and can work with any width.