Remove predefined blob samples
Before, sample selections of blobs were hand-picked.

Now that we have a dynamic list of servers from the API, we actually end
up loading blobs *twice*: the hand-picked ones, then ones that were
randomly picked from the API response.

Here we remove the hand-picked blobs to 1) avoid bias 2) avoid loading
blobs twice if the user has JavaScript enabled (and the API stuff kicks
in), which will be most of the time.

This ends speeding up loading a little.