Debounce search processing
If a user types "thi" into the search field, the component will perform
the filtering operation three times. For just "t", around 1,000 results
will be returned, but never actually shown. (It's only shown for the
brief moment that only the "t" is on screen.)

This patch adds debounced search processing as long as 3 or less
characters are in the search field.