Represent internal structures with neater classes
Instead of keeping everything as plain objects and arrays, let's wrap
them all in classes that make the huge tree of data more organized. This
will make it easier to implement fine-grained filtering in the future
without clutter.

Also, multiple internal instances of "server" were replaced with
"guild" to be consistent with how the Discord API works. I might extend
this naming scheme to the rest of the codebase in the future.